Separate the features API from the frontend app
