+++ packages/wordpress-playground/src/features/login.ts
@@ -1,0 +1,21 @@
+import type { PlaygroundAPI } from "../app";
+
+export async function login(
+	playground: PlaygroundAPI,
+	user = 'admin',
+	password = 'password'
+) {
+	await playground.request({
+		relativeUrl: '/wp-login.php',
+	});
+
+	await playground.request({
+		relativeUrl: '/wp-login.php',
+		method: 'POST',
+		formData: {
+			log: user,
+			pwd: password,
+			rememberme: 'forever',
+		}
+	});
+}
