+++ packages/wordpress-playground/src/wp-client.ts
@@ -1,195 +1,10 @@
-import { PHPResponse } from '@wordpress/php-wasm';
 import type { PlaygroundAPI } from './app';
+import { consumeAPI } from '@wordpress/php-wasm'
 
-export async function login(
-	playground: PlaygroundAPI,
-	user = 'admin',
-	password = 'password'
-) {
-	await playground.request({
-		relativeUrl: '/wp-login.php',
-	});
-
-	await playground.request({
-		relativeUrl: '/wp-login.php',
-		method: 'POST',
-		formData: {
-			log: user,
-			pwd: password,
-			rememberme: 'forever',
-		}
-	});
-}
-
-export async function installPlugin(
-	playground: PlaygroundAPI,
-	pluginZipFile: File,
-	options: any = {}
-) {
-	const activate = 'activate' in options ? options.activate : true;
-
-	// Upload it to WordPress
-	const pluginForm = await playground.request({
-		relativeUrl: '/wp-admin/plugin-install.php?tab=upload',
-	});
-	const pluginFormPage = asDOM(pluginForm);
-	const pluginFormData = new FormData(
-		pluginFormPage.querySelector('.wp-upload-form')! as HTMLFormElement
-	) as any;
-	const { pluginzip, ...postData } = Object.fromEntries(
-		pluginFormData.entries()
-	);
-
-	const pluginInstalledResponse = await playground.request({
-		relativeUrl: '/wp-admin/update.php?action=upload-plugin',
-		method: 'POST',
-		formData: postData,
-		files: { pluginzip: pluginZipFile },
-	});
-
-	// Activate if needed
-	if (activate) {
-		const pluginInstalledPage = asDOM(pluginInstalledResponse);
-		const activateButtonHref = pluginInstalledPage
-			.querySelector('#wpbody-content .button.button-primary')!
-			.attributes.getNamedItem('href')!.value;
-		const activatePluginUrl = new URL(
-			activateButtonHref,
-			await playground.pathToInternalUrl('/wp-admin/')
-		).toString();
-		await playground.request({
-			absoluteUrl: activatePluginUrl,
-		});
+export function connectToPlayground(iframe: HTMLIFrameElement, url: string) {
+	if (!url.endsWith('/wordpress.html')) {
+		url = url.replace(/\/+$/, '') + '/wordpress.html';
 	}
-
-	/**
-	 * Pair the site editor's nested iframe to the Service Worker.
-	 *
-	 * Without the patch below, the site editor initiates network requests that
-	 * aren't routed through the service worker. That's a known browser issue:
-	 *
-	 * * https://bugs.chromium.org/p/chromium/issues/detail?id=880768
-	 * * https://bugzilla.mozilla.org/show_bug.cgi?id=1293277
-	 * * https://github.com/w3c/ServiceWorker/issues/765
-	 *
-	 * The problem with iframes using srcDoc and src="about:blank" as they
-	 * fail to inherit the root site's service worker.
-	 *
-	 * Gutenberg loads the site editor using <iframe srcDoc="<!doctype html">
-	 * to force the standards mode and not the quirks mode:
-	 *
-	 * https://github.com/WordPress/gutenberg/pull/38855
-	 *
-	 * This commit patches the site editor to achieve the same result via
-	 * <iframe src="/doctype.html"> and a doctype.html file containing just
-	 * `<!doctype html>`. This allows the iframe to inherit the service worker
-	 * and correctly load all the css, js, fonts, images, and other assets.
-	 *
-	 * Ideally this issue would be fixed directly in Gutenberg and the patch
-	 * below would be removed.
-	 *
-	 * See https://github.com/WordPress/wordpress-playground/issues/42 for more details
-	 */
-	if (
-		(await playground.isDir('/wordpress/wp-content/plugins/gutenberg')) &&
-		!(await playground.fileExists('/wordpress/.gutenberg-patched'))
-	) {
-		async function patchFile(path, callback) {
-			return await playground.writeFile(
-				path,
-				callback(await playground.readFileAsText(path))
-			);
-		}
-		await playground.writeFile('/wordpress/.gutenberg-patched', '1');
-		await patchFile(
-			`/wordpress/wp-content/plugins/gutenberg/build/block-editor/index.js`,
-			(contents) =>
-				contents.replace(
-					/srcDoc:("[^"]+"|[^,]+)/g,
-					'src:"/wp-includes/empty.html"'
-				)
-		);
-		await patchFile(
-			`/wordpress/wp-content/plugins/gutenberg/build/block-editor/index.min.js`,
-			(contents) =>
-				contents.replace(
-					/srcDoc:("[^"]+"|[^,]+)/g,
-					'src:"/wp-includes/empty.html"'
-				)
-		);
-	}
-}
-
-export async function installTheme(
-	playground: PlaygroundAPI,
-	themeZipFile: File,
-	options: any = {}
-) {
-	const activate = 'activate' in options ? options.activate : true;
-
-	// Upload it to WordPress
-	const themeForm = await playground.request({
-		relativeUrl: '/wp-admin/theme-install.php'
-	});
-	const themeFormPage = asDOM(themeForm);
-	const themeFormData = new FormData(
-		themeFormPage.querySelector('.wp-upload-form')! as HTMLFormElement
-	) as any;
-	const { themezip, ...postData } = Object.fromEntries(
-		themeFormData.entries()
-	);
-
-	const themeInstalledResponse = await playground.request({
-		relativeUrl: '/wp-admin/update.php?action=upload-theme',
-		method: 'POST',
-		formData: postData,
-		files: { themezip: themeZipFile },
-	});
-
-	// Activate if needed
-	if (activate) {
-		const themeInstalledPage = asDOM(themeInstalledResponse);
-
-		const messageContainer = themeInstalledPage.querySelector(
-			'#wpbody-content > .wrap'
-		);
-		if (
-			messageContainer?.textContent?.includes(
-				'Theme installation failed.'
-			)
-		) {
-			console.error(messageContainer?.textContent);
-			return;
-		}
-
-		const activateButton = themeInstalledPage.querySelector(
-			'#wpbody-content .activatelink, ' +
-				'.update-from-upload-actions .button.button-primary'
-		);
-		if (!activateButton) {
-			console.error('The "activate" button was not found.');
-			return;
-		}
-
-		const activateButtonHref =
-			activateButton.attributes.getNamedItem('href')!.value;
-		const activateThemeUrl = new URL(
-			activateButtonHref,
-			await playground.pathToInternalUrl('/wp-admin/')
-		).toString();
-		await playground.request({
-			absoluteUrl: activateThemeUrl,
-		});
-	}
-}
-
-function asDOM(response: PHPResponse) {
-	return new DOMParser().parseFromString(
-		asText(response),
-		'text/html'
-	)!;
-}
-
-function asText(response: PHPResponse) {
-	return new TextDecoder().decode(response.body);
+	iframe.src = url;
+	return consumeAPI(iframe.contentWindow!) as PlaygroundAPI;
 }
